cleanup: Convert all time parameters to signed
This converts all time parameters to signed, finishing a cleanup that
was started in 2018 in b4256bf0.  This changes the type of:
  - timestamps
    - PTS specifically
  - timestamp offsets
  - timescales
  - durations

This excludes:
  - MP4 box definitions
  - DTS specifically

This is meant to address signed/unsigned conversion issues on arm64
that caused some test cases to fail.

+++ packager/media/formats/mp4/fragmenter.cc
@@ -90,8 +90,7 @@ Status Fragmenter::AddSample(const MediaSample& sample) {
 
   if (stream_info_->stream_type() == StreamType::kStreamVideo &&
       sample.is_key_frame()) {
-    key_frame_infos_.push_back(
-        {static_cast<uint64_t>(pts), data_->Size(), sample.data_size()});
+    key_frame_infos_.push_back({pts, data_->Size(), sample.data_size()});
   }
 
   data_->AppendArray(sample.data(), sample.data_size());
